Pull podcast information from RSS feed
Pull title, description, image and url from a provided rss feed url in
data. For the initial sorting a title has to be provided, hence title is
required. Even when a RSS feed url is provided, the properties can be
overridden by listing the property in the rssOverride array, e.g.:

{
  // title is needed for sorting, but title from RSS feed will be used
  title: 'Some Title',
  // description from RSS feed is too long so it should be overridden
  rssOverride: ['description'],
  // so this will be displayed because of the rssOverride entry
  description: 'Some short description',
  // url will only be used, if the RSS feed does not contain a link
  url: 'https://someurl.dev',
  // url to the RSS feed
  rss: 'https://someurl.dev/rss',
}